+++ catsup/generator/renderer.py
@@ -34,6 +34,7 @@ class Renderer(object):
         theme_filter_path = os.path.join(g.theme.path, 'filters.py')
         self.load_filters_from_pyfile(catsup_filter_path)
         self.load_filters_from_pyfile(theme_filter_path)
+        self.rendered_permalinks = []
 
     def load_filters_from_pyfile(self, path):
         if not os.path.exists(path):
@@ -50,16 +51,23 @@ class Renderer(object):
             pass
 
     def render_to(self, template, permalink, **kwargs):
-        kwargs.setdefault("permalink", urljoin(
+        permalink, output_name = urljoin(
             g.base_url,
             permalink
-        ))
+        ), permalink
+        kwargs.setdefault("permalink", permalink)
         html = self.render(template, **kwargs)
         if html:
             output_name = permalink
-            if output_name.endswith("/"):
-                output_name += 'index.html'
+            self.rendered_permalinks.append(permalink)
+            if output_name.endswith("/") or "." not in output_name:
+                output_name = output_name.rstrip("/")
+                output_name += '/index.html'
             output_path = os.path.join(g.output, output_name.lstrip("/"))
             mkdir(os.path.dirname(output_path))
             with open(output_path, "w") as f:
                 f.write(html)
+
+    def render_sitemap(self):
+        with open(os.path.join(g.output, "sitemap.txt"), "w") as f:
+            f.write("\n".join(self.rendered_permalinks))